+++ test/resources/formatting_test_fprettify_long_lines.f90
@@ -1,0 +1,6 @@
+program demo
+
+    write(*, "('Just a very long line. Just a very long line. Just a very long line. Just a very long line. Just a very long line. Just a very long line. ',i0)") 100
+ 
+endprogram
+